+++ lib/gutenberg.php
@@ -9,7 +9,7 @@
  * @since 	1.0.2
  * @license GPL-2.0+
  * @link    http://studiomoare.com/
- * @version 1.0.2
+ * @version 1.0.3
  */
 
 add_action( 'after_setup_theme', 'mg_setup_theme_supported_features' );